+++ jfconnect/manager.go
@@ -4,11 +4,10 @@ import (
 	"github.com/jfrog/jfrog-client-go/config"
 	"github.com/jfrog/jfrog-client-go/http/jfroghttpclient"
 	"github.com/jfrog/jfrog-client-go/jfconnect/services"
-	"github.com/jfrog/jfrog-client-go/jfconnect/services/metrics"
 )
 
 type Manager interface {
-	PostVisibilityMetric(metrics.VisibilityMetric) error
+	PostVisibilityMetric(services.VisibilityMetric) error
 }
 
 type jfConnectManager struct {
@@ -40,7 +39,7 @@ func (jm *jfConnectManager) Client() *jfroghttpclient.JfrogHttpClient {
 	return jm.client
 }
 
-func (jm *jfConnectManager) PostVisibilityMetric(metric metrics.VisibilityMetric) error {
+func (jm *jfConnectManager) PostVisibilityMetric(metric services.VisibilityMetric) error {
 	jfConnectService := services.NewJfConnectService(jm.config.GetServiceDetails(), jm.client)
 	return jfConnectService.PostVisibilityMetric(metric)
 }
